+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix problem with Date field in User app reports does not populate when a first entry was cleared. Refs UIU-2991.
 * Hide all actionalble buttons on user details pane for DCB Virtual user. Refs UIU-2987.
 * Open loan page modifications for a virtual patron. Refs UIU-2988.
+* Fix wrong date in Cash-Drawer-Reconciliation-Report.pdf. Refs UIU-3010.
 
 ## [10.0.4](https://github.com/folio-org/ui-users/tree/v10.0.4) (2023-11-10)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v10.0.3...v10.0.4)